p011 - Interface changes
A few changes to the interface have been made. Now you have a new set of
parameters that influence the curve derivative calculation (which is a
key step in finding the contact point and in finding the peaks) in the
"Curve Params" tab. The parameters for the peaks finding algorithm nave
been moved to the "Peaks" tab.